factor out encodeKey to package since it is shared by several modules
