+++ spark/src/main/scala/geotrellis/spark/io/s3/spatial/SpatialRasterRDDWriter.scala
@@ -18,9 +18,10 @@ import com.amazonaws.services.s3.model.AmazonS3Exception
 import scala.reflect.ClassTag
 
 object SpatialRasterRDDWriter extends RasterRDDWriter[SpatialKey] with LazyLogging {
-  val encodeKey = (key: SpatialKey, ki: KeyIndex[SpatialKey], max: Int) =>
+  val encodeKey = (key: SpatialKey, ki: KeyIndex[SpatialKey], max: Int) => {
     ki.toIndex(key).toString.reverse.padTo(max, '0').reverse
-
+  }
+  
   def getKeyBounds(rdd: RasterRDD[SpatialKey]): KeyBounds[SpatialKey] = {
     val md = rdd.metaData
     val gb = md.gridBounds
