+++ spark/src/main/scala/geotrellis/spark/io/s3/spatial/package.scala
@@ -1,0 +1,13 @@
+package geotrellis.spark.io.s3
+
+import geotrellis.spark._
+import geotrellis.spark.io.index._
+import org.joda.time.format.ISODateTimeFormat
+
+package object spatial {
+
+  private[spatial]
+  val encodeKey = (key: SpatialKey, ki: KeyIndex[SpatialKey], max: Int) => {
+    ki.toIndex(key).toString.reverse.padTo(max, '0').reverse
+  }
+}
